ref: Move the colibri queue from ConferenceShim to Conference.
