+++ jvb/src/main/java/org/jitsi/videobridge/Videobridge.java
@@ -414,7 +414,7 @@ public class Videobridge
         }
 
         // It is now the responsibility of Conference to send a response.
-        conference.getShim().enqueueColibriRequest(request);
+        conference.enqueueColibriRequest(request);
     }
 
     private @NotNull Conference getOrCreateConference(ColibriConferenceIQ conferenceIq)
