+++ jvb/src/main/java/org/jitsi/videobridge/shim/ConferenceShim.java
@@ -20,11 +20,9 @@ import org.jitsi.nlj.format.*;
 import org.jitsi.nlj.rtp.*;
 import org.jitsi.utils.*;
 import org.jitsi.utils.logging2.*;
-import org.jitsi.utils.queue.*;
 import org.jitsi.videobridge.*;
 import org.jitsi.videobridge.octo.*;
 import org.jitsi.videobridge.util.*;
-import org.jitsi.videobridge.xmpp.*;
 import org.jitsi.xmpp.extensions.colibri.*;
 import org.jitsi.xmpp.extensions.jingle.*;
 import org.jitsi.xmpp.util.*;
@@ -60,8 +58,6 @@ public class ConferenceShim
      */
     private final Map<MediaType, ContentShim> contents = new HashMap<>();
 
-    private final PacketQueue<XmppConnection.ColibriRequest> colibriQueue;
-
     /**
      * Initializes a new {@link ConferenceShim} instance.
      *
@@ -71,48 +67,7 @@ public class ConferenceShim
     {
         this.logger = parentLogger.createChildLogger(ConferenceShim.class.getName());
         this.conference = conference;
-        colibriQueue = new PacketQueue<>(
-                Integer.MAX_VALUE,
-                true,
-                "colibri-queue",
-                request ->
-                {
-                    try
-                    {
-                        long start = System.currentTimeMillis();
-                        IQ response = handleColibriConferenceIQ(request.getRequest());
-                        long end = System.currentTimeMillis();
-                        long processingDelay = end - start;
-                        long totalDelay = end - request.getReceiveTime();
-                        request.getProcessingDelayStats().addDelay(processingDelay);
-                        request.getTotalDelayStats().addDelay(totalDelay);
-                        if (processingDelay > 100)
-                        {
-                            logger.warn("Took " + processingDelay + " ms to process an IQ (total delay "
-                                    + totalDelay + " ms): " + request.getRequest().toXML());
-                        }
-                        request.getCallback().invoke(response);
-                    }
-                    catch (Throwable e)
-                    {
-                        logger.warn("Failed to handle colibri request: ", e);
-                        request.getCallback().invoke(
-                                IQUtils.createError(
-                                        request.getRequest(),
-                                        StanzaError.Condition.internal_server_error,
-                                        e.getMessage()));
-                    }
-                    return true;
-                },
-                TaskPools.IO_POOL
-        );
     }
-
-    public void enqueueColibriRequest(XmppConnection.ColibriRequest request)
-    {
-        colibriQueue.add(request);
-    }
-
 
     /**
      * Gets the content of type {@code type}, creating it if necessary.
@@ -378,11 +333,6 @@ public class ConferenceShim
                 endpoint.setStatsId(colibriEndpoint.getStatsId());
             }
         }
-    }
-
-    public void close()
-    {
-        colibriQueue.close();
     }
 
     /**
